change set function
w.getAngle() function was used in set() function.
This function accesseed to the angle of joint with std::string,
but set function used jointID(enum) to detect the joint name before.
I added the part of switching with jointID, so now we can use set
function with jointID and it can get current joint angle with
std::string