fix the set function
There is a bug in set() function, movementbase.cpp.
The local variable in set() which is called jointName, is the array
contains the names of joint.
It is exsisted to convert enum jointID to std::string,
but the order of jointName wasn't correspond with jointID's.
Now this bug is fixed.