Changing the arguments of some functions.
I changed the arguments of some functions.
Those functions are getNextAngle(), renew(), and setAngle()
that are in the files of Movement class in hillstoneunited/movement/.

For getNextAngle(), I added the third arguments int& ts to calcurate t
from ts(timestamp).

For renew(), I changed argument to int& ts from Action& act. The reason
of this change is as same as above.

For setAngle(), I erased the first argument, Action& act. It's because
this argument was unnecessary.