making ffwalk
This is the half way of my work,
so this program haven't generate walking behavior yet.

now I making the function the name of which is inverseKinematics().
This function will calculate the angles of each joint in the low half of
the body from the position where we want the agent's tips of the toe move to.